backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
feedparser==6.0.10
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
openai==1.3.7